clean up login & signup
