+++ src/components/SignupSuccess.jsx
@@ -1,0 +1,5 @@
+const SignupSuccess = () => {
+  return <div>REGISTRATION COMPLETED</div>;
+};
+
+export default SignupSuccess;
