+++ src/pages/LogIn.jsx
@@ -1,5 +1,38 @@
-const LogIn = () => {
-  return <h1>Login page</h1>;
-};
+import { useState } from 'react';
 
-export default LogIn;
+import LoginForm from '../components/LoginForm';
+
+function Login() {
+  const adminUser = {
+    email: 'test@email.com',
+    password: 'test123',
+  };
+
+  const [user, setUser] = useState({
+    email: '',
+  });
+
+  const [error, setError] = useState('');
+
+  const LoginDetails = (details) => {
+    if (details.email === adminUser.email && details.password === adminUser.password) {
+      setUser({
+        email: details.email,
+      });
+    } else {
+      setError('Invalid email or password!');
+    }
+  };
+
+  return (
+    <div className="App">
+      {user.email !== '' ? (
+        <div>{(window.location.pathname = '/try-it-out')}</div>
+      ) : (
+        <LoginForm Login={LoginDetails} error={error} />
+      )}
+    </div>
+  );
+}
+
+export default Login;
