--- src/pages/Register/SignupSucces.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const SignupSuccess = () => {
-  return (
-      <div>
-          REGISTRATION COMPLETED
-      </div>
-  )
-};
-
-export default SignupSuccess;
